moved classes to lib/dispatcher.py
moved SdiOutputDispatcher, NotificationDispatcher, and VidOutputDispatcher to lib/dispatcher.py. layer.py
